Run clusterregistrationtoken controllers is prov v2 or mcm is enabled
